requirements.txt
# Core dependencies
moviepy==2.2.1
openai-whisper==20231117
spellchecker==0.8.3

# Audio processing
pydub==0.25.1

# HTTP requests for translation API
requests==2.32.4

# Data processing
numpy==2.2.6

# Utilities
tqdm==4.67.1
python-dotenv==1.1.1
 